:sparkles: feat: API 모듈 변경 및 axios 설정 업데이트
- 기존 api.ts 파일을 삭제하고 axios.ts 파일로 대체하여 API 호출을 위한 모듈을 통합하였습니다.
- 요청 인터셉터를 통해 accessToken을 자동으로 헤더에 첨부하도록 구현하였으며, 응답 인터셉터의 기본 구조를 추가하였습니다.
- 인증 API 호출 시 새로운 axios 모듈을 사용하도록 변경하여 코드의 일관성을 높였습니다.
